Refactor file methods
