+++ src/main/java/seedu/programmer/ui/MainWindow.java
@@ -23,6 +23,8 @@ import javafx.stage.Stage;
 import seedu.programmer.commons.core.GuiSettings;
 import seedu.programmer.commons.core.LogsCenter;
 import seedu.programmer.commons.exceptions.IllegalValueException;
+import seedu.programmer.commons.util.FileUtil;
+import seedu.programmer.commons.util.JsonUtil;
 import seedu.programmer.logic.Logic;
 import seedu.programmer.logic.commands.CommandResult;
 import seedu.programmer.logic.commands.DashboardCommandResult;
@@ -33,7 +35,6 @@ import seedu.programmer.logic.commands.ShowCommandResult;
 import seedu.programmer.logic.commands.UploadCommandResult;
 import seedu.programmer.logic.commands.exceptions.CommandException;
 import seedu.programmer.logic.parser.exceptions.ParseException;
-import seedu.programmer.model.FileManager;
 import seedu.programmer.model.ProgrammerError;
 import seedu.programmer.model.student.Student;
 import seedu.programmer.model.student.exceptions.DuplicateStudentException;
@@ -227,7 +228,6 @@ public class MainWindow extends UiPart<Stage> {
      */
     @FXML
     private void handleUpload() {
-        FileManager fm = new FileManager();
         File chosenFile = promptUserForCsvFile();
         if (chosenFile == null) {
             logger.info("User cancelled the file upload.");
@@ -236,7 +236,7 @@ public class MainWindow extends UiPart<Stage> {
 
         List<Student> stuList;
         try {
-            stuList = fm.getStudentsFromCsv(chosenFile);
+            stuList = FileUtil.getStudentsFromCsv(chosenFile);
         } catch (IllegalArgumentException | IOException e) {
             displayPopup("Upload failed: " + e.getMessage()); // Error with file data
             return;
@@ -263,8 +263,7 @@ public class MainWindow extends UiPart<Stage> {
      */
     @FXML
     private void handleDownload() {
-        FileManager fm = new FileManager();
-        JSONArray jsonData = fm.getJsonData("data/programmerError.json");
+        JSONArray jsonData = JsonUtil.getJsonData("data/programmerError.json");
         assert (jsonData != null);
 
         if (jsonData.length() == 0) {
@@ -274,7 +273,7 @@ public class MainWindow extends UiPart<Stage> {
 
         File destinationFile = promptUserForDestination();
         if (destinationFile != null) {
-            fm.writeJsonToCsv(jsonData, destinationFile);
+            JsonUtil.writeJsonToCsv(jsonData, destinationFile);
             displayPopup("Your data has been downloaded to " + destinationFile + "!");
             logger.info("Data successfully downloaded as CSV.");
         }
@@ -290,6 +289,7 @@ public class MainWindow extends UiPart<Stage> {
         assert (message != null);
         Popup popup = createPopup(message);
         double tenPercent = 1 - NINETY_PERCENT;
+
         // Add some left padding according to primaryStage's width
         popup.setX(primaryStage.getX() + primaryStage.getWidth() * tenPercent / 2);
 
@@ -358,16 +358,6 @@ public class MainWindow extends UiPart<Stage> {
         return chosenDir == null ? null : new File(chosenDir, destFileName);
     }
 
-    @FXML
-    private void handleHover() {
-        exitButton.setStyle("-fx-background-color: -fx-light-bg-color;");
-    }
-
-    @FXML
-    private void handleUnhover() {
-        exitButton.setStyle("-fx-background-color: -fx-main-bg-color;");
-    }
-
     /**
      * Executes the command and returns the result.
      *
@@ -399,5 +389,15 @@ public class MainWindow extends UiPart<Stage> {
             resultDisplay.setFeedbackToUser(e.getMessage());
             throw e;
         }
+    }
+
+    @FXML
+    private void handleHover() {
+        exitButton.setStyle("-fx-background-color: -fx-light-bg-color;");
+    }
+
+    @FXML
+    private void handleUnhover() {
+        exitButton.setStyle("-fx-background-color: -fx-main-bg-color;");
     }
 }
