+++ src/test/java/seedu/programmer/commons/util/JsonUtilTest.java
@@ -1,10 +1,12 @@
 package seedu.programmer.commons.util;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 import java.io.IOException;
 import java.nio.file.Path;
 
+import org.json.JSONArray;
 import org.junit.jupiter.api.Test;
 
 import seedu.programmer.testutil.SerializableTestClass;
@@ -15,7 +17,11 @@ import seedu.programmer.testutil.TestUtil;
  */
 public class JsonUtilTest {
 
+    private static final String TEST_DATA_FOLDER = "src/test/data/JsonUtilTest/";
     private static final Path SERIALIZATION_FILE = TestUtil.getFilePathInSandboxFolder("serialize.json");
+    private static final String VALID_JSON = TEST_DATA_FOLDER + "programmerError.json";
+    private static final String INVALID_JSON = TEST_DATA_FOLDER + "invalidProgrammerError.json";
+    private static final int DEFAULT_NUM_OF_STUDENTS = 36;
 
     @Test
     public void serializeObjectToJsonFile_noExceptionThrown() throws IOException {
@@ -39,7 +45,16 @@ public class JsonUtilTest {
         assertEquals(serializableTestClass.getMapOfIntegerToString(), SerializableTestClass.getHashMapTestValues());
     }
 
-    //TODO: @Test jsonUtil_readJsonStringToObjectInstance_correctObject()
+    @Test
+    public void getJsonData_validJson_returnsCorrectNumberOfElements() {
+        JSONArray result = JsonUtil.getJsonData(VALID_JSON);
+        assert result != null;
+        assertEquals(result.length(), DEFAULT_NUM_OF_STUDENTS);
+    }
 
-    //TODO: @Test jsonUtil_writeThenReadObjectToJson_correctObject()
+    @Test
+    public void getJsonData_invalidJson_returnsNull() {
+        JSONArray result = JsonUtil.getJsonData(INVALID_JSON);
+        assertNull(result);
+    }
 }
